refactor: make controllers adapter-agnostic (no AR exception references)
Phase 3 of Mongoid adapter implementation:
- Replace ActiveRecord::Rollback with IronAdmin::Rollback in controller
- Replace ActiveRecord::RecordNotFound with IronAdmin::RecordNotFound
- Replace record.saved_changes with adapter.record_changes(record)
- Use adapter.wrap_rollback for transaction rollback handling
- Replace record_scope.find with adapter-agnostic find_record helper
- Extract Scopeable concern (base_scope, record_scope, find_record)
