+++ spec/integration/edge_cases_spec.rb
@@ -34,26 +34,26 @@ RSpec.describe "Edge Cases", type: :request do
     it "raises error for non-existent record on show" do
       expect do
         get iron_admin.resource_path("users", 99_999)
-      end.to raise_error(ActiveRecord::RecordNotFound)
+      end.to raise_error(IronAdmin::RecordNotFound)
     end
 
     it "raises error for non-existent record on edit" do
       expect do
         get iron_admin.edit_resource_path("users", 99_999)
-      end.to raise_error(ActiveRecord::RecordNotFound)
+      end.to raise_error(IronAdmin::RecordNotFound)
     end
 
     it "raises error for non-existent record on update" do
       expect do
         patch iron_admin.resource_path("users", 99_999),
               params: { record: { name: "Test" } }
-      end.to raise_error(ActiveRecord::RecordNotFound)
+      end.to raise_error(IronAdmin::RecordNotFound)
     end
 
     it "raises error for non-existent record on destroy" do
       expect do
         delete iron_admin.resource_path("users", 99_999)
-      end.to raise_error(ActiveRecord::RecordNotFound)
+      end.to raise_error(IronAdmin::RecordNotFound)
     end
   end
